+++ agent_structs/structs_task_messages.go
@@ -3,14 +3,23 @@ package agentstructs
 // PT_TASK_* structs
 
 type PTTaskMessageAllData struct {
-	Task            PTTaskMessageTaskData                `json:"task"`
-	Callback        PTTaskMessageCallbackData            `json:"callback"`
+	// Task - Read-only data about the task
+	Task PTTaskMessageTaskData `json:"task"`
+	// Callback - Read-only data about the callback
+	Callback PTTaskMessageCallbackData `json:"callback"`
+	// BuildParameters - Read-only data about the build parameters
 	BuildParameters []PayloadConfigurationBuildParameter `json:"build_parameters"`
-	Commands        []string                             `json:"commands"`
-	Payload         PTTaskMessagePayloadData             `json:"payload"`
-	C2Profiles      []PayloadConfigurationC2Profile      `json:"c2info"`
-	PayloadType     string                               `json:"payload_type"`
-	Args            PTTaskMessageArgsData
+	// Commands - Read-only data about the commands built into the callback
+	Commands []string `json:"commands"`
+	// Payload - Read-only data about the backing payload for this task
+	Payload PTTaskMessagePayloadData `json:"payload"`
+	// C2Profiles - Read-only data about the c2 profiles and their values for this callback
+	C2Profiles []PayloadConfigurationC2Profile `json:"c2info"`
+	// PayloadType - Read-only the name of the payload type associated with this task
+	PayloadType string `json:"payload_type"`
+	// Args - Read-Write argument data for adding/removing/modifying args associated with this task instance.
+	// Mainly for create tasking function to augment parameters
+	Args PTTaskMessageArgsData
 }
 
 type PTTaskMessageTaskData struct {
@@ -87,6 +96,9 @@ type PT_TASK_FUNCTION_STATUS = string
 
 type PtTaskFunctionParseArgString func(args *PTTaskMessageArgsData, input string) error
 type PtTaskFunctionParseArgDictionary func(args *PTTaskMessageArgsData, input map[string]interface{}) error
+
+// PTTaskMessageArgsData - struct for tracking, adding, removing, updating, validating, etc arguments for a task.
+// If you want to set your own manual arguments, use the .SetManualArgs function.
 type PTTaskMessageArgsData struct {
 	args            []CommandParameter
 	commandLine     string
@@ -128,7 +140,7 @@ const (
 	OPSEC_ROLE_OTHER_OPERATOR            = "other_operator"
 )
 
-type PtTaskFunctionOPSECPre func(PTTaskMessageAllData) PTTTaskOPSECPreTaskMessageResponse
+type PtTaskFunctionOPSECPre func(*PTTaskMessageAllData) PTTTaskOPSECPreTaskMessageResponse
 type PTTTaskOPSECPreTaskMessageResponse struct {
 	TaskID             int        `json:"task_id"`
 	Success            bool       `json:"success"`
@@ -141,30 +153,42 @@ type PTTTaskOPSECPreTaskMessageResponse struct {
 
 // Tasking step 2:
 // Task message/process to run the create_tasking function
-//
-//	this can start creating subtasks
-type PtTaskFunctionCreateTasking func(PTTaskMessageAllData) PTTaskCreateTaskingMessageResponse
+
+// PtTaskFunctionCreateTasking - Process the tasking request from the user. If you want to access/modify the arguments
+// for this task, use the Task.Args.* functions.
+type PtTaskFunctionCreateTasking func(*PTTaskMessageAllData) PTTaskCreateTaskingMessageResponse
 type PTTaskCreateTaskingMessageResponse struct {
-	TaskID                 int     `json:"task_id"`
-	Success                bool    `json:"success"`
-	Error                  string  `json:"error"`
-	CommandName            *string `json:"command_name,omitempty"`
-	TaskStatus             *string `json:"task_status,omitempty"`
-	DisplayParams          *string `json:"display_params,omitempty"`
-	Stdout                 *string `json:"stdout,omitempty"`
-	Stderr                 *string `json:"stderr,omitempty"`
-	Completed              *bool   `json:"completed,omitempty"`
-	TokenID                *int    `json:"token_id,omitempty"`
+	// TaskID - the task associated with the create tasking function - this will be automatically filled in for you
+	TaskID int `json:"task_id"`
+	// Success - indicate if the create tasking function ran successfully or not
+	Success bool `json:"success"`
+	// Error - if you want to provide an error message about some error you hit while executing the create tasking
+	Error string `json:"error"`
+	// CommandName - if you want to change the associated command name that's sent down to the agent
+	CommandName *string `json:"command_name,omitempty"`
+	// TaskStatus - if you want to manually set the task status to be something other than default
+	TaskStatus *string `json:"task_status,omitempty"`
+	// DisplayParams - if you want to change the display parameters for your task to be something other than the default JSON
+	DisplayParams *string `json:"display_params,omitempty"`
+	// Stdout - Provide any task-based stdout
+	Stdout *string `json:"stdout,omitempty"`
+	// Stderr - Provide any task-based stderr
+	Stderr *string `json:"stderr,omitempty"`
+	// Completed - identify if the task is already completed and shouldn't be sent down to the agent
+	Completed *bool `json:"completed,omitempty"`
+	// TokenID - identifier for the token id associated with this task - normally doesn't need to be set unless you're changing it
+	TokenID *uint64 `json:"token_id,omitempty"`
+	// CompletionFunctionName - name of the completion function to call from the Command's TaskCompletionFunctions dictionary
 	CompletionFunctionName *string `json:"completion_function_name,omitempty"`
-	Params                 string  `json:"params"`
-	ParameterGroupName     string  `json:"parameter_group_name"`
+	// ParameterGroupName - manually give the parameter group name instead of having Mythic determine it
+	ParameterGroupName string `json:"parameter_group_name"`
 }
 
 // Tasking step 3:
 // Task message/process after running create_tasking but before the task can be picked up by an agent
 //
 //	this is the time to check any artifacts generated from create_tasking
-type PtTaskFunctionOPSECPost func(PTTaskMessageAllData) PTTaskOPSECPostTaskMessageResponse
+type PtTaskFunctionOPSECPost func(*PTTaskMessageAllData) PTTaskOPSECPostTaskMessageResponse
 type PTTaskOPSECPostTaskMessageResponse struct {
 	TaskID              int        `json:"task_id"`
 	Success             bool       `json:"success"`
@@ -180,12 +204,12 @@ type PTTaskOPSECPostTaskMessageResponse struct {
 type SubtaskGroupName = string
 
 type PTTaskCompletionFunctionMessage struct {
-	TaskData               PTTaskMessageAllData  `json:"task"`
+	TaskData               *PTTaskMessageAllData `json:"task"`
 	SubtaskData            *PTTaskMessageAllData `json:"subtask,omitempty"`
 	SubtaskGroup           *SubtaskGroupName     `json:"subtask_group_name,omitempty"`
 	CompletionFunctionName string                `json:"function_name"`
 }
-type PTTaskCompletionFunction func(PTTaskMessageAllData, *PTTaskMessageAllData, *SubtaskGroupName) PTTaskCompletionFunctionMessageResponse
+type PTTaskCompletionFunction func(*PTTaskMessageAllData, *PTTaskMessageAllData, *SubtaskGroupName) PTTaskCompletionFunctionMessageResponse
 type PTTaskCompletionFunctionMessageResponse struct {
 	TaskID                 int     `json:"task_id"`
 	ParentTaskId           int     `json:"parent_task_id"`
@@ -205,8 +229,8 @@ type PTTaskCompletionFunctionMessageResponse struct {
 // Tasking step 5:
 // Task message/process to run for more manual processing of a message's response data
 type PtTaskProcessResponseMessage struct {
-	TaskData PTTaskMessageAllData `json:"task"`
-	Response interface{}          `json:"response"`
+	TaskData *PTTaskMessageAllData `json:"task"`
+	Response interface{}           `json:"response"`
 }
 type PtTaskFunctionProcessResponse func(PtTaskProcessResponseMessage) PTTaskProcessResponseMessageResponse
 type PTTaskProcessResponseMessageResponse struct {