+++ CHANGELOG.md
@@ -4,6 +4,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v1.0.0-0.0.9] - 2023-03-15
+
+### Changed
+
+- Updated create tasking functions to take pointer rather than value
+- Started adding text descriptions for structs to make it easier for development
+
+## [v1.0.0-0.0.8] - 2023-03-14
+
+### Changed
+
+- updated some structs to uint64 from int to match Mythic
+
+## [v1.0.0-0.0.7] - 2023-03-12
+
+### Changed
+
+- fixed an issue with the Process response message routing to itself
+
 ## [v1.0.0-0.0.6] - 2023-03-03
 
 ### Changed